Rewrite Scripting bot respecting the Bot API + Code Optimization
Do NOT put thread sleeps in bots. NEVER. Initialize() and Update() are
called in the main thread!
Also, a delay should be respected between each chat/command sendind to
avoid being kicked.
Putting Thread.Sleep() in bots may cause the client to be out of sync
with the server or make it lose connection so again do not use them :)