Added a new PM regex
Added
[someone @ me] message
so that remote control could be used on more servers.

Added back Hero Chat Messages with a default config option of;
herochatmessagesenabled=false
which safely disables it for everyone and has to be explicitly enabled
for it to goof anything up.
How this happened was me downloading the "source" from minecraft forums
and modifying that source then trying to merge it back to the main
github fork of mine.