Check player is in range each entity attack
This commit should make it so that MCC will check that the player is in range before attacking entities each iteration of the cooldown.

This was problematic before, because, if the entities were still they would be added to the attack list, and then if they stayed still and the client teleported away, it would still attempt to attack the entities, even though it is now out of range.