Allow specifying authlib-injector API path
The previous authlib-injector authentication implementation always
assumed that the authlib-injector API location was located at
https://authlibinjectorserver.example.com/api/yggdrasil. Blessing Skin
structures their API like that, but other authlib-injector-compatible
Yggdrasil implementations do not. Per the authlib-injector
specification, the API root can be located at any path, and that path
should be pointed to by the `X-Authlib-Injector-API-Location` header
[1].

With this change, the default AuthlibInjectorAPIPath is kept as
`/api/yggdrasil`, so users will not need to update their config.

See also
https://github.com/unmojang/drasl/issues/66#issuecomment-2081817705.

[1] https://github.com/yushijinhun/authlib-injector/wiki/Yggdrasil-%E6%9C%8D%E5%8A%A1%E7%AB%AF%E6%8A%80%E6%9C%AF%E8%A7%84%E8%8C%83