Fetch certificates for Yggdrasil accounts
Some authlib-injector-compatible authentication servers implement the
POST /player/certificates route used for fetching player certificates.

To comply with the authlib-injector Yggdrasil server specification [1], we
first query the authlib-injector metadata and check for the
`feature.enable_profile_key` flag. If the flag is set, the
authentication server supports the /player/certificates route.

In this case, checking the authlib-injector metadata isn't really
necessary; if the /player/certificates request fails for any reason, we
simply ignore it anyway.

[1] https://github.com/yushijinhun/authlib-injector/wiki/Yggdrasil-%E6%9C%8D%E5%8A%A1%E7%AB%AF%E6%8A%80%E6%9C%AF%E8%A7%84%E8%8C%83